+++ volmdlr/faces.py
@@ -33,6 +33,7 @@ import volmdlr.wires
 import volmdlr.display as vmd
 import volmdlr.geometry
 import volmdlr.grid
+import volmdlr.utils.parametric as vm_parametric
 
 
 def knots_vector_inv(knots_vector):
@@ -45,117 +46,6 @@ def knots_vector_inv(knots_vector):
     multiplicities = [knots_vector.count(knot) for knot in knots]
 
     return knots, multiplicities
-
-
-def repair_singularity(primitive, last_primitive):
-    """
-    Repairs the Contour2D of SphericalSurface3D and ConicalSurface3D parametric face representations.
-
-    Used when transforming from spatial to parametric coordinates when the surface contains a sigularity
-    """
-    v1 = primitive.unit_direction_vector()
-    v2 = last_primitive.unit_direction_vector()
-    dot = v1.dot(volmdlr.X2D)
-    cross = v1.cross(v2)
-    new_primitives = []
-    if cross == 0 and dot == 0:
-        if primitive.start.x == math.pi:
-            primitive = primitive.translation(volmdlr.Vector2D(-2 * math.pi, 0))
-            new = vme.LineSegment2D(last_primitive.end, primitive.start)
-        elif primitive.start.x == -math.pi:
-            primitive = primitive.translation(volmdlr.Vector2D(2 * math.pi, 0))
-            new = vme.LineSegment2D(last_primitive.end, primitive.start)
-        else:
-            new = vme.LineSegment2D(last_primitive.end, primitive.start)
-
-        new_primitives.append(new)
-        new_primitives.append(primitive)
-    else:
-        delta = last_primitive.end - primitive.start
-        new_primitives.append(primitive.translation(delta))
-    return new_primitives
-
-
-def repair_start_end_angle_periodicity(angle_start, angle_end, ref_start, ref_end):
-    """
-    Repairs start and end angles in parametric coordinates based in ref_start (angle just after start_angle)
-     and ref_end (angle just before angle_end).
-    """
-    # Verify if theta1 or theta2 point should be -pi because atan2() -> ]-pi, pi]
-    if math.isclose(angle_start, math.pi, abs_tol=1e-6) and ref_start < 0:
-        angle_start = -math.pi
-    elif math.isclose(angle_start, -math.pi, abs_tol=1e-6) and ref_start > 0:
-        angle_start = math.pi
-    if math.isclose(angle_end, math.pi, abs_tol=1e-6) and ref_end < 0:
-        angle_end = -math.pi
-    elif math.isclose(angle_end, -math.pi, abs_tol=1e-6) and ref_end > 0:
-        angle_end = math.pi
-    return angle_start, angle_end
-
-
-def repair_arc3d_angle_continuity(angle_start, angle_after_start, angle_end, angle3d):
-    """
-    Repairs Arc3D angle continuity on parametric 2D space.
-    """
-    ref_low = angle_start - angle3d
-    ref_up = angle_start + angle3d
-    # angle_after_start < angle_start --> angle coordinate going clockwise
-    # ref_low < -math.pi -> passing lower bound (-math.pi)
-    if angle_after_start < angle_start and ref_low < -math.pi:
-        angle_end = ref_low
-    # angle_after_start > angle_start --> angle coordinate going trigowise
-    #  ref_up > math.pi -> passing upper bound (math.pi)
-    elif angle_after_start > angle_start and ref_up > math.pi:
-        angle_end = ref_up
-
-    if angle_start > 0 > angle_after_start:
-        angle_start -= 2 * math.pi
-    elif angle_start < 0 < angle_after_start:
-        angle_start += 2 * math.pi
-
-    return angle_start, angle_end
-
-
-def arc3d_to_cylindrical_verification(start, end, angle3d, theta3, theta4):
-    """
-    Verifies theta from start and end of an arc3d after transformation from spatial to parametric coordinates.
-    """
-    theta1, z1 = start
-    theta2, z2 = end
-
-    theta1, theta2 = repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
-
-    theta1, theta2 = repair_arc3d_angle_continuity(theta1, theta3, theta2, angle3d)
-
-    start = volmdlr.Point2D(theta1, z1)
-    end = volmdlr.Point2D(theta2, z2)
-    return [start, end]
-
-
-def arc3d_to_spherical_verification(start, end, angle3d, point_after_start, point_before_end):
-    """
-    Verifies theta and phi from start and end of an arc3d after transformation from spatial to parametric coordinates.
-    """
-    theta1, phi1 = start
-    theta2, phi2 = end
-    theta3, phi3 = point_after_start
-    theta4, phi4 = point_before_end
-    # Verify if theta1 or theta2 point should be -pi or pi because atan2() -> ]-pi, pi]
-    theta1, theta2 = repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
-
-    # Verify if phi1 or phi2 point should be -pi or pi because phi -> ]-pi, pi]
-    phi1, phi2 = repair_start_end_angle_periodicity(phi1, phi2, phi3, phi4)
-
-    if math.isclose(phi1, phi2, abs_tol=1e-4):
-        theta1, theta2 = repair_arc3d_angle_continuity(theta1, theta3, theta2, angle3d)
-
-    if math.isclose(theta1, theta2, abs_tol=1e-4):
-        phi1, phi2 = repair_arc3d_angle_continuity(phi1, phi3, phi2, angle3d)
-
-    start = volmdlr.Point2D(theta1, phi1)
-    end = volmdlr.Point2D(theta2, phi2)
-
-    return start, end
 
 
 class Surface2D(volmdlr.core.Primitive2D):
@@ -1451,7 +1341,7 @@ class CylindricalSurface3D(Surface3D):
         theta3, _ = self.point3d_to_2d(arc3d.point_at_abscissa(0.001 * length))
         theta4, _ = self.point3d_to_2d(arc3d.point_at_abscissa(0.98 * length))
 
-        start, end = arc3d_to_cylindrical_verification(start, end, angle3d, theta3, theta4)
+        start, end = vm_parametric.arc3d_to_cylindrical_verification(start, end, angle3d, theta3, theta4)
 
         return [vme.LineSegment2D(start, end)]
 
@@ -1523,7 +1413,7 @@ class CylindricalSurface3D(Surface3D):
         theta4, _ = self.point3d_to_2d(arcellipse3d.point_at_abscissa(0.98 * length))
 
         # Verify if theta1 or theta2 point should be -pi because atan2() -> ]-pi, pi]
-        theta1, theta2 = repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
+        theta1, theta2 = vm_parametric.repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
 
         points[0] = volmdlr.Point2D(theta1, z1)
         points[-1] = volmdlr.Point2D(theta2, z2)
@@ -1564,7 +1454,7 @@ class CylindricalSurface3D(Surface3D):
         theta4, _ = self.point3d_to_2d(bspline_curve3d.point_at_abscissa(0.98 * length))
 
         # Verify if theta1 or theta2 point should be -pi because atan2() -> ]-pi, pi]
-        theta1, theta2 = repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
+        theta1, theta2 = vm_parametric.repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
 
         points[0] = volmdlr.Point2D(theta1, z1)
         points[-1] = volmdlr.Point2D(theta2, z2)
@@ -2025,7 +1915,7 @@ class ToroidalSurface3D(Surface3D):
         point_after_start = self.point3d_to_2d(arc3d.point_at_abscissa(0.001 * length))
         point_before_end = self.point3d_to_2d(arc3d.point_at_abscissa(0.98 * length))
 
-        start, end = arc3d_to_spherical_verification(start, end, angle3d, point_after_start,
+        start, end = vm_parametric.arc3d_to_spherical_verification(start, end, angle3d, point_after_start,
                                                      point_before_end)
 
         return [vme.LineSegment2D(start, end)]
@@ -2042,10 +1932,10 @@ class ToroidalSurface3D(Surface3D):
         points = [self.point3d_to_2d(p) for p in bspline_curve3d.discretization_points(number_points=11)]
 
         # Verify if theta1 or theta2 point should be -pi because atan2() -> ]-pi, pi]
-        theta1, theta2 = repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
+        theta1, theta2 = vm_parametric.repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
 
         # Verify if phi1 or phi2 point should be -pi because phi -> ]-pi, pi]
-        phi1, phi2 = repair_start_end_angle_periodicity(phi1, phi2, phi3, phi4)
+        phi1, phi2 = vm_parametric.repair_start_end_angle_periodicity(phi1, phi2, phi3, phi4)
 
         points[0] = volmdlr.Point2D(theta1, phi1)
         points[-1] = volmdlr.Point2D(theta2, phi2)
@@ -2079,7 +1969,7 @@ class ToroidalSurface3D(Surface3D):
         theta4, _ = points[-2]
 
         # Verify if theta1 or theta2 point should be -pi because atan2() -> ]-pi, pi]
-        theta1, theta2 = repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
+        theta1, theta2 = vm_parametric.repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
 
         points[0] = volmdlr.Point2D(theta1, z1)
         points[-1] = volmdlr.Point2D(theta2, z2)
@@ -2257,7 +2147,7 @@ class ConicalSurface3D(Surface3D):
         theta3, _ = self.point3d_to_2d(arc3d.point_at_abscissa(0.001 * length))
         theta4, _ = self.point3d_to_2d(arc3d.point_at_abscissa(0.98 * length))
 
-        start, end = arc3d_to_cylindrical_verification(start, end, angle3d, theta3, theta4)
+        start, end = vm_parametric.arc3d_to_cylindrical_verification(start, end, angle3d, theta3, theta4)
 
         return [vme.LineSegment2D(start, end)]
 
@@ -2276,7 +2166,7 @@ class ConicalSurface3D(Surface3D):
         theta4, _ = self.point3d_to_2d(bspline_curve3d.point_at_abscissa(0.98 * length))
 
         # Verify if theta1 or theta2 point should be -pi because atan2() -> ]-pi, pi]
-        theta1, theta2 = repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
+        theta1, theta2 = vm_parametric.repair_start_end_angle_periodicity(theta1, theta2, theta3, theta4)
 
         points[0] = volmdlr.Point2D(theta1, z1)
         points[-1] = volmdlr.Point2D(theta2, z2)
